refactor: move trap logic to shared
