+++ Content.Client/_Scp/Other/AirlockManEater/AirlockManEaterSystem.cs
@@ -1,0 +1,8 @@
+﻿using Content.Shared._Scp.Other.AirlockManEater;
+
+namespace Content.Client._Scp.Other.AirlockManEater;
+
+public sealed class AirlockManEaterSystem : SharedAirlockManEaterSystem
+{
+
+}
